fix creation of vmi file
When we where writing in a vmi file after the actual EOF, it did not work.
Using open instead of fopen which works.

It did not work because with fopen, in "w" or "w+" mode, it truncate the file,
so we lost every data previously written.
And in "a" mode, the writing position is always the EOF:

Opening a file in append mode (a as the first character of mode) causes all
subsequent write operations to this stream to occur at end-of-file, as if
preceded the call:
fseek(stream, 0, SEEK_END);